rag-embedding-service/requirements.txt
# Django and web framework
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
requests==2.32.3

# Redis
redis==6.2.0
hiredis==2.4.0

# LangChain and related dependencies
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
openai>=1.86.0,<2.0.0

# Document processing
pypdf==5.9.0
python-docx==1.1.2

# Vector database
pgvector==0.4.1

# Other utilities
numpy==1.26.4
tiktoken==0.9.0
python-magic==0.4.27
python-multipart==0.0.9

# WebSocket support
channels==4.3.0
channels-redis==4.3.0
daphne==4.2.1

